Move filepath logic "up"
Previously the GetUniqueEnvironmentFolder function took care of joining
the repoPath with the environments folder so that it could use the
DirectoriesUnderPath function. But callers outside of the package could
not do that, so I don't think they could call DirectoriesUnderPath
properly.

This moves the handling of the repoPath into the DirectoriesUnderPath
function so that outside callers can use it.