Add options for from-branch and to-branch (#106)
* Add toBranch and fromBranch

The new `fromBranch` and `toBranch` determine which branches to use for
the new service configuration, and where to base the Pull Request.
Previously, both of these were hard-coded to "master".

I think there is some confusion in naming here... the Promote function
already takes an (undocumented) `newBranchName` parameter, which is
later just called `branchName` in the makePullRequestInput function.
That's just the name of the (temporary) branch created on the
destination repo, where the changes are pushed so they can be merged
into the "Base" branch (previously hard-coded as "master", now the new
`toBranch` parameter).

Fixes: #100

* Add EnvLocation and refactor

EnvLocation holds information about the environment's location so that
it can be passed around more compactly in functions. Currently it only
holds the Path (URL or local) and the Branch, but it could hold other
information, for example the environment folder in Repositories with
multiple environments.