Add sub-commands for specific promotion paths
These new sub-commands simplify the interface when promoting between
branches in a repository, between environments in a repository, or
between two "simple" repositories. The previous behavior is preserved if
no sub-command is given: the user can specify any combination of
repository, branch and environment folder (even a local repo).

For #92